refactor(cli): extract shared trust-root utilities module (#176)
