+++ crates/pi-coding-agent/src/skills_commands.rs
@@ -1,64 +1,5 @@
 use super::*;
 
-pub(crate) fn parse_trust_rotation_spec(raw: &str) -> Result<(String, TrustedKey)> {
-    let (old_id, new_spec) = raw.split_once(':').ok_or_else(|| {
-        anyhow!("invalid --skill-trust-rotate '{raw}', expected old_id:new_id=base64_key")
-    })?;
-    let old_id = old_id.trim();
-    if old_id.is_empty() {
-        bail!("invalid --skill-trust-rotate '{raw}', expected old_id:new_id=base64_key");
-    }
-    let new_key = parse_trusted_root_spec(new_spec)?;
-    Ok((old_id.to_string(), new_key))
-}
-
-pub(crate) fn load_trust_root_records(path: &Path) -> Result<Vec<TrustedRootRecord>> {
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    let raw = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read {}", path.display()))?;
-    let parsed = serde_json::from_str::<TrustedRootFileFormat>(&raw)
-        .with_context(|| format!("failed to parse trusted root file {}", path.display()))?;
-
-    let records = match parsed {
-        TrustedRootFileFormat::List(items) => items,
-        TrustedRootFileFormat::Wrapped { roots } => roots,
-        TrustedRootFileFormat::Keys { keys } => keys,
-    };
-
-    Ok(records)
-}
-
-pub(crate) fn save_trust_root_records(path: &Path, records: &[TrustedRootRecord]) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        if !parent.as_os_str().is_empty() {
-            std::fs::create_dir_all(parent)
-                .with_context(|| format!("failed to create {}", parent.display()))?;
-        }
-    }
-    let mut payload = serde_json::to_string_pretty(&TrustedRootFileFormat::Wrapped {
-        roots: records.to_vec(),
-    })
-    .context("failed to serialize trusted root records")?;
-    payload.push('\n');
-    write_text_atomic(path, &payload)
-        .with_context(|| format!("failed to write {}", path.display()))?;
-    Ok(())
-}
-
-pub(crate) fn apply_trust_root_mutations(
-    records: &mut Vec<TrustedRootRecord>,
-    cli: &Cli,
-) -> Result<TrustMutationReport> {
-    apply_trust_root_mutation_specs(
-        records,
-        &cli.skill_trust_add,
-        &cli.skill_trust_revoke,
-        &cli.skill_trust_rotate,
-    )
-}
-
 pub(crate) struct SkillsSearchMatch {
     pub(crate) name: String,
     pub(crate) file: String,
